requirements.txt
Django>=1.11,<2.0
psycopg2
git+https://github.com/jonhillmtl/django-jwt-auth
git+https://github.com/jonhillmtl/jwt-key-gen
djangorestframework
djangorestframework-jsonapi